Make HierarchyRendererCellReferenceWrapper a nested class of HierarchyRenderer
